Move src/legacy/server/keystore to src/cli.
Currently keystore parsing is in the legacy http server folder.
Keystore references end in src/cli so instead of migrating to
core/server the I opted for the closest fit in the CLI folder.

Closes #52107
