+++ src/cli/keystore/errors.js
