+++ src/cli/keystore/index.js
