+++ src/cli/keystore/keystore.js
