+++ src/cli/keystore/keystore.test.js
