+++ src/cli/serve/serve.js
@@ -27,7 +27,7 @@ import { getConfigPath } from '@kbn/utils';
 import { IS_KIBANA_DISTRIBUTABLE } from '../../legacy/utils';
 import { fromRoot } from '../../core/server/utils';
 import { bootstrap } from '../../core/server';
-import { readKeystore } from './read_keystore';
+import { readKeystore } from '../keystore/read_keystore';
 
 function canRequire(path) {
   try {
