+++ src/cli_keystore/add.test.js
@@ -39,7 +39,7 @@ jest.mock('fs', () => ({
 import sinon from 'sinon';
 import { PassThrough } from 'stream';
 
-import { Keystore } from '../legacy/server/keystore';
+import { Keystore } from '../cli/keystore';
 import { add } from './add';
 import { Logger } from '../cli_plugin/lib/logger';
 import * as prompt from './utils/prompt';
