+++ src/cli_keystore/list.test.js
@@ -36,7 +36,7 @@ jest.mock('fs', () => ({
 }));
 
 import sinon from 'sinon';
-import { Keystore } from '../legacy/server/keystore';
+import { Keystore } from '../cli//keystore';
 import { list } from './list';
 import { Logger } from '../cli_plugin/lib/logger';
